method for distribution blocks, move randomStr to Utils #8
